+++ datafusion/core/src/execution/memory_pool/proxy.rs
