+++ datafusion/core/src/execution/runtime_env.rs
@@ -27,7 +27,7 @@ use std::collections::HashMap;
 use crate::datasource::datasource::TableProviderFactory;
 use crate::datasource::listing_table_factory::ListingTableFactory;
 use crate::datasource::object_store::ObjectStoreRegistry;
-use crate::execution::memory_manager::{
+use crate::execution::memory_pool::{
     GreedyMemoryPool, MemoryPool, UnboundedMemoryPool,
 };
 use datafusion_common::DataFusionError;
